+++ ilert/data_source_connection.go
@@ -1,0 +1,65 @@
+package ilert
+
+import (
+	"fmt"
+	"log"
+	"time"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/iLert/ilert-go"
+)
+
+func dataSourceConnection() *schema.Resource {
+	return &schema.Resource{
+		Read: dataSourceConnectionRead,
+
+		Schema: map[string]*schema.Schema{
+			"name": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"trigger_mode": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+		},
+	}
+}
+
+func dataSourceConnectionRead(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*ilert.Client)
+
+	log.Printf("[DEBUG] Reading iLert connection")
+
+	searchName := d.Get("name").(string)
+
+	return resource.Retry(2*time.Minute, func() *resource.RetryError {
+		resp, err := client.GetConnections(&ilert.GetConnectionsInput{})
+		if err != nil {
+			time.Sleep(2 * time.Second)
+			return resource.RetryableError(err)
+		}
+
+		var found *ilert.ConnectionOutput
+
+		for _, connection := range resp.Connections {
+			if connection.Name == searchName {
+				found = connection
+				break
+			}
+		}
+
+		if found == nil {
+			return resource.NonRetryableError(
+				fmt.Errorf("Unable to locate any connection with the name: %s", searchName),
+			)
+		}
+
+		d.SetId(found.ID)
+		d.Set("name", found.Name)
+		d.Set("trigger_mode", found.TriggerMode)
+
+		return nil
+	})
+}