+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## (Unreleased), Version 1.0.1
+## (Unreleased), Version 1.1.0
 
+- add connection data source and resource
+- add connector data source
 - remove type argument from escalation rule
 - user standard user agent header for each request